Fix LU-15: Broken ROI image
This fix contains two changes: 1. We check that passed ROI requests are
multiples of possible ROI steps as provided by the camera. If this is not the
case, the request is ignored and a warning issued. 2. We added two new base
properties ROI_WIDTH_MULTIPLIER and ROI_HEIGHT_MULTIPLIER that expose this
information to client programs.