Fix #69: revert back to fixed mock camera settings
The combination of CONSTRUCT_ONLY and READWRITE causes a) unnecessary warnings
and b) confuses the TANGO server down below (see SCHNEIDE-661). Since adapting
the mock camera size and bitdepth at run-time is not a high-priority feature, we
reverted that change. In the future and if still required this should be solved
through either compile-time defines or an environment variable.